+++ src/Math/include/BipedalLocomotion/Math/Spline.h
@@ -71,7 +71,7 @@ public:
 
     /**
      * Set the knots of the spline.
-     * @param position position of the knots in \$f\mathbb{R}^n\$f.
+     * @param position position of the knots in \f$\mathbb{R}^n\f$.
      * @param time vector containing the time instant of the knots.
      * @return True in case of success, false otherwise.
      */