+++ src/Math/tests/SplineTest.cpp
@@ -13,6 +13,7 @@
 #include <catch2/catch_test_macros.hpp>
 
 #include <BipedalLocomotion/Math/CubicSpline.h>
+#include <BipedalLocomotion/Math/LinearSpline.h>
 #include <BipedalLocomotion/Math/QuinticSpline.h>
 
 using namespace BipedalLocomotion::Math;
@@ -306,6 +307,115 @@ TEST_CASE("Cubic spline")
             expected = coefficients[1] //
                        + 2 * coefficients[2] * t //
                        + 3 * coefficients[3] * std::pow(t, 2);
+            REQUIRE(expected.isApprox(velocityVector[i], 1e-5));
+        }
+    }
+}
+
+TEST_CASE("Linear spline")
+{
+    using namespace std::chrono_literals;
+
+    constexpr std::size_t numberOfPoints = 100;
+    constexpr std::chrono::nanoseconds initTime = 320ms;
+    constexpr std::chrono::nanoseconds finalTime = 2s + 640s;
+
+    constexpr std::chrono::nanoseconds dT = (finalTime - initTime) / (numberOfPoints - 1);
+    std::array<Eigen::Vector4d, 2> coefficients;
+    for (auto& coeff : coefficients)
+    {
+        coeff.setRandom();
+    }
+
+    std::vector<Eigen::Vector4d> knots;
+    std::vector<std::chrono::nanoseconds> time;
+    for (std::size_t i = 0; i < numberOfPoints; i++)
+    {
+        time.push_back(dT * i + initTime);
+        const double t = std::chrono::duration<double>(time.back()).count();
+        knots.push_back(coefficients[0] + coefficients[1] * t);
+    }
+
+    const double initT = std::chrono::duration<double>(initTime).count();
+    const double finalT = std::chrono::duration<double>(finalTime).count();
+
+    Eigen::Vector4d initVelocity = coefficients[1];
+    Eigen::Vector4d finalVelocity = coefficients[1];
+
+    Eigen::Vector4d initAcceleration = Eigen::Vector4d::Zero();
+    Eigen::Vector4d finalAcceleration = Eigen::Vector4d::Zero();
+
+    LinearSpline<Eigen::Vector4d> spline;
+    REQUIRE(spline.setKnots(knots, time));
+
+    REQUIRE(spline.setInitialConditions({initVelocity, initAcceleration}));
+    REQUIRE(spline.setFinalConditions({finalVelocity, finalAcceleration}));
+
+    constexpr std::size_t pointsToCheckNumber = 1e3;
+
+    constexpr std::chrono::nanoseconds dTCheckPoints = (finalTime - initTime) / pointsToCheckNumber;
+
+    Eigen::Vector4d expected, position, velocity, acceleration;
+
+    for (std::size_t i = 0; i < pointsToCheckNumber; i++)
+    {
+        std::chrono::nanoseconds t = dTCheckPoints * i + initTime;
+        double tSeconds = std::chrono::duration<double>(t).count();
+
+        REQUIRE(spline.evaluatePoint(t, position, velocity, acceleration));
+
+        // check position
+        expected = coefficients[0] + coefficients[1] * tSeconds;
+
+        REQUIRE(expected.isApprox(position, 1e-5));
+    }
+
+    SECTION("Advance capabilities")
+    {
+        REQUIRE_FALSE(spline.isOutputValid());
+        REQUIRE(spline.setAdvanceTimeStep(dTCheckPoints));
+
+        for (std::size_t i = 0; i < pointsToCheckNumber; i++)
+        {
+            double t = std::chrono::duration<double>(dTCheckPoints * i + initTime).count();
+
+            // advance the spline
+            REQUIRE(spline.advance());
+            REQUIRE(spline.isOutputValid());
+            const auto& traj = spline.getOutput();
+
+            // check position
+            expected = coefficients[0] + coefficients[1] * t;
+
+            REQUIRE(expected.isApprox(traj.position, 1e-5));
+        }
+    }
+
+    SECTION("Query from a vector of times")
+    {
+        std::vector<std::chrono::nanoseconds> timeVector;
+        for (std::size_t i = 0; i < pointsToCheckNumber; i++)
+        {
+            timeVector.push_back(dTCheckPoints * i + initTime);
+        }
+
+        std::vector<Eigen::Vector4d> positionVector, velocityVector, accelerationVector;
+        REQUIRE(spline.evaluateOrderedPoints(timeVector, //
+                                             positionVector,
+                                             velocityVector,
+                                             accelerationVector));
+
+        for (std::size_t i = 0; i < pointsToCheckNumber; i++)
+        {
+            double t = std::chrono::duration<double>(dTCheckPoints * i + initTime).count();
+
+            // check position
+            expected = coefficients[0] + coefficients[1] * t;
+
+            REQUIRE(expected.isApprox(positionVector[i], 1e-5));
+
+            // check velocity
+            expected = coefficients[1];
             REQUIRE(expected.isApprox(velocityVector[i], 1e-5));
         }
     }